Introduce directory for factories and move  modeldocuments there
