+++ source/libmvvm_model/mvvm/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_subdirectory(commands)
 add_subdirectory(core)
+add_subdirectory(factories)
 add_subdirectory(interfaces)
 add_subdirectory(model)
 add_subdirectory(project)
