+++ source/libmvvm_model/mvvm/core/CMakeLists.txt
@@ -2,8 +2,6 @@ target_sources(mvvm_model PRIVATE
     export.h
     filesystem.h
     modeldocumentinterface.h
-    modeldocuments.cpp
-    modeldocuments.h
     types.h
     uniqueidgenerator.cpp
     uniqueidgenerator.h
