+++ source/libmvvm_model/mvvm/factories/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(mvvm_model PRIVATE
+    modeldocuments.cpp
+    modeldocuments.h
+)
